refactored Output to generate dynamic code, delete unused files
