+++ client/components/Output.tsx
@@ -2,50 +2,42 @@ import React, {useContext} from 'react';
 import DefaultForm from './DefaultForm';
 import { CopyBlock, dracula, nord, monokai } from 'react-code-blocks';
 import ScrollToBottom from 'react-scroll-to-bottom';
-import TestRenderer from 'react-test-renderer';
-import Navbar from './NavBar';
 import { Resizable } from 're-resizable';
-// import jsxToString from 'jsx-to-string';
-// import ContextProvider from '../../src/ContextProvider';
-import ContextProvider from '../containers/ContextProvider';
+import { AppContext } from '../../src/'
 
 export default function Output() {
-  const state = useContext(ContextProvider);
-  console.log(state); // an array of the components
-
-  const injectFunc = (state) => {
+  /// use hook to check the current elements dropped on canva 
+  const { listOfDroppedElements }: any = useContext(AppContext);
+  // get all output strings in an array
+  const injectFunc = (listofDraggableElements:any) => {
     const result = [];
-    for(let obj of state.componentOptions) {
-      // console.log(obj)
+    for(let obj of listOfDroppedElements) {
       result.push(obj.data);
     }
     return result
   }
 
-  const componentResults = injectFunc(state);
+
+  //display each output string on a separate line in right order
+  const componentResults = injectFunc(listOfDroppedElements);
   let cleanedUpResults = ``;
   for (const each of componentResults) {
     cleanedUpResults += `\n\t` + each;
   } 
 
-  const parse =  `
-  import React from "react";
-  import { useForm } from "react-hook-form";
+  const parse = `import React from "react";
+import { useForm } from "react-hook-form";
   
-  type Inputs = {
-    example: string,
-    exampleRequired: string,
-  };
-  
-  export default function DefaultForm() {
-    const { register, handleSubmit, watch, errors } = useForm<Inputs>();
+
+  export default function App() {
+    const { register, handleSubmit, watch, errors } = useForm();
+    const onSubmit = data => console.log(data);
+
     return (
       <div className = 'default-form'>${cleanedUpResults}
       </div>
     );
-  };
-      
-      `
+  };`
 
   return (
     <Resizable
