--- client/components/Test.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-export default function Test() {
-  return (
-   <div>
-     <input></input>
-     <button>Submit</button>
-   </div> 
-  );
-}
